project ready for deployment

+++ my-springApp/src/main/java/net/airlineSystem/my_springApp/repositery/BookingDetailsDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import net.airlineSystem.my_springApp.model.BookingDetails;
 
-public interface BookingDetailsDao extends JpaRepository<BookingDetails, Integer> {
+public interface BookingDetailsDao extends JpaRepository<BookingDetails, Long> {
  
 }
